posts/requirements.txt
fastapi[standard]>=0.115.0
pydantic>=2.9.0
PyJWT>=2.9.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
